% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{package}
\name{aviculture-package}
\alias{aviculture}
\alias{aviculture-package}
\title{Aviculture Growth Model for Ross 308 Broilers}
\description{
An R package for simulating and predicting broiler chicken growth curves
based on a multi-tissue density model. The package implements interpolation
methods (Spline, Lagrange, Gompertz, RBF) to model mass and volume growth
dynamics.
}
\examples{
\dontrun{
  # Load the pre-trained model
  model <- load_model()

  # Get mass at age 30 days
  mass_30 <- predict_mass(30, model)
  print(mass_30)

  # Get growth rate
  gamma_m <- predict_growth_rate(30, mass_30, model)
  print(gamma_m)
}

}
\seealso{
\code{\link{load_model}}, \code{\link{predict_mass}},
\code{\link{predict_volume}}, \code{\link{predict_growth_rate}}
}
\keyword{internal}
