% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{compute_indicators}
\alias{compute_indicators}
\title{Compute simulation indicators}
\usage{
compute_indicators(simulation)
}
\arguments{
\item{simulation}{An object created by \code{\link[=solve_aviculture]{solve_aviculture()}}.}
}
\value{
A data frame of time-indexed indicators.
}
\description{
Returns the indicator table stored in a simulation object.
}
