% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{compute_objectives}
\alias{compute_objectives}
\title{Compute objective values}
\usage{
compute_objectives(simulation)
}
\arguments{
\item{simulation}{An object created by \code{\link[=solve_aviculture]{solve_aviculture()}}.}
}
\value{
A named list with \code{J1} and \code{J2}.
}
\description{
Returns the objective values corresponding to the simulated trajectory.
}
