% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{create_controls}
\alias{create_controls}
\title{Create exogenous controls}
\usage{
create_controls(
  grid = create_grid(),
  supply = NULL,
  demand = NULL,
  price = NULL,
  supply_cost = NULL,
  name = "default"
)
}
\arguments{
\item{grid}{A grid created with \code{\link[=create_grid]{create_grid()}}.}

\item{supply}{Chicks supply rate. Can be \code{NULL}, a scalar, a vector defined
on \code{grid$time}, or a function of time.}

\item{demand}{Market demand. Can be \code{NULL}, a scalar, a vector defined on
\code{grid$time}, or a function of time.}

\item{price}{Selling price. Can be \code{NULL}, a scalar, a vector defined on
\code{grid$time}, or a function of time.}

\item{supply_cost}{Unit supply cost. Can be \code{NULL}, a scalar, a vector
defined on \code{grid$time}, or a function of time.}

\item{name}{Optional descriptive label.}
}
\value{
An object of class \code{aviculture_controls}.
}
\description{
Defines time-dependent supply, selling price, and demand controls. Scalars
and vectors are automatically converted to time functions.
}
