% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{get_model_info}
\alias{get_model_info}
\title{Model Metadata and Quality Information}
\usage{
get_model_info(model, verbose = FALSE)
}
\arguments{
\item{model}{A model object from \code{\link{load_model}}.}

\item{verbose}{Logical. If TRUE, print formatted output. Default is FALSE.}
}
\value{
A list containing:
\item{model_name}{Interpolation method used}
\item{training_ages}{Age range of training data}
\item{n_samples}{Number of training samples}
\item{parameters}{Model parameters (alpha, beta)}
\item{quality_metrics}{RMSE, stability, convergence}
}
\description{
Returns information about the loaded model including interpolation method,
training data, and quality metrics.
}
\examples{
\dontrun{
  model <- load_model()
  info <- get_model_info(model, verbose = TRUE)
}

}
