% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{get_tissue_densities}
\alias{get_tissue_densities}
\title{Tissue Density Reference Values}
\usage{
get_tissue_densities(model, tissue = NULL)
}
\arguments{
\item{model}{A model object from \code{\link{load_model}}.}

\item{tissue}{Character. Name of tissue: "bone", "muscle", "fat", "blood", "organs".
If NULL, returns all tissues.}
}
\value{
If tissue is NULL: a named list of tissue densities (kg/L).
If tissue is specified: the density value (numeric) for that tissue.
}
\description{
Returns the reference density values for different broiler chicken tissues.
}
\details{
Tissue densities are biological constants determined from literature:
\itemize{
\item bone: 1.85 kg/L (mineralized bone)
\item muscle: 1.06 kg/L (protein + water)
\item fat: 0.90 kg/L (lipids)
\item blood: 1.06 kg/L (similar to water)
\item organs: 1.04 kg/L (viscera)
}
}
\examples{
\dontrun{
  model <- load_model()
  
  # Get all tissue densities
  densities <- get_tissue_densities(model)
  
  # Get specific tissue
  bone_density <- get_tissue_densities(model, tissue = "bone")
}

}
