\name{aviculture_simulation_methods}
\alias{print.aviculture_grid}
\alias{print.aviculture_parameters}
\alias{print.aviculture_controls}
\alias{print.aviculture_simulation}
\alias{summary.aviculture_simulation}
\alias{plot.aviculture_simulation}
\title{S3 methods for aviculture objects}
\description{
Print, summary, and default plot methods for the core aviculture object
classes.
}
