% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{plot_sensitivity}
\alias{plot_sensitivity}
\title{Plot sensitivity results}
\usage{
plot_sensitivity(sensitivity, metric = "J1", overlay = FALSE)
}
\arguments{
\item{sensitivity}{An object created by \code{\link[=run_sensitivity]{run_sensitivity()}}.}

\item{metric}{Metric to display.}

\item{overlay}{Logical; if \code{TRUE}, plots all variables on the same figure.}
}
\value{
Invisibly returns the plotting data.
}
\description{
Produces two-dimensional sensitivity plots for a selected metric. Curves can
be shown together or one by one.
}
